[F]MOD:换了一个视频，修改了video页面进不去的bug
