+++ src/main/java/controller/MainPageController.java
@@ -1,4 +1,44 @@
 package controller;
 
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.geometry.Pos;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.Slider;
+import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.media.Media;
+import javafx.scene.media.MediaPlayer;
+import javafx.scene.media.MediaView;
+import javafx.stage.Stage;
+import javafx.util.Duration;
+
+import java.io.IOException;
+
 public class MainPageController {
+    public javafx.scene.control.Button test;
+    @FXML
+    public void initialize() {
+         test.setVisible(false);
+    }
+
+    public void showVideo(MouseEvent mouseEvent) throws IOException {
+        Stage stage = (Stage) test.getScene().getWindow();
+        stage.setTitle("Video Show");
+        FXMLLoader loader = new FXMLLoader();
+        loader.setLocation(getClass().getResource("/view/fxml/VideoShow.fxml"));
+        AnchorPane video = (AnchorPane) loader.load();
+        // Set person overview into the center of root layout.
+        AnchorPane anchorPane= (AnchorPane) stage.getScene().getRoot();
+        anchorPane.getChildren().remove(2);
+        anchorPane.getChildren().add(2, video);
+
+        video.setLayoutX(200);
+        video.setLayoutY(75);
+
+    }
 }
