+++ src/main/resources/view/fxml/VideoShow.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.media.*?>
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<AnchorPane prefHeight="675" prefWidth="1000.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.VideoShowController">
+   <children>
+      <MediaView fx:id="mView" fitHeight="600.0" fitWidth="800.0" layoutX="100.0" layoutY="14.0" />
+   </children>
+</AnchorPane>
